Feature/create middleware

+++ SelXPressApi/SelXPressApi/Controllers/AttributeDataController.cs
@@ -4,7 +4,7 @@ using SelXPressApi.DocumentationErrorTemplate;
 using SelXPressApi.DTO.AttributeDataDTO;
 using SelXPressApi.Exceptions;
 using SelXPressApi.Interfaces;
-using SelXPressApi.Models;
+using SelXPressApi.Middleware;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
@@ -19,16 +19,21 @@ namespace SelXPressApi.Controllers
     {
         private readonly IAttributeDataRepository _attributeDataRepository;
         private readonly IMapper _mapper;
+        private readonly IAuthorizationMiddleware _authorizationMiddleware;
 
-        public AttributeDataController(IAttributeDataRepository attributeDataRepository, IMapper mapper)
+        public AttributeDataController(IAttributeDataRepository attributeDataRepository, IMapper mapper, IAuthorizationMiddleware authorizationMiddleware)
         {
             _attributeDataRepository = attributeDataRepository;
             _mapper = mapper;
+            _authorizationMiddleware = authorizationMiddleware;
         }
 
         /// <summary>
-        /// Get all AttributeData items.
+        /// Method to get all the attribute data in the database
         /// </summary>
+        /// <returns></returns>
+        /// <exception cref="BadRequestException"></exception>
+        /// <exception cref="NotFoundException"></exception>
         [HttpGet]
         [ProducesResponseType(200, Type = typeof(List<AttributeDataDto>))]
         [ProducesResponseType(404, Type = typeof(NotFoundErrorTemplate))]
@@ -48,8 +53,12 @@ namespace SelXPressApi.Controllers
         }
 
         /// <summary>
-        /// Get an AttributeData item by ID.
+        /// Method to get the attribute data based on the id
         /// </summary>
+        /// <param name="id"></param>
+        /// <returns></returns>
+        /// <exception cref="NotFoundException"></exception>
+        /// <exception cref="BadRequestException"></exception>
         [HttpGet("{id}")]
         [ProducesResponseType(200, Type = typeof(AttributeDataDto))]
         [ProducesResponseType(400, Type = typeof(BadRequestErrorTemplate))]
@@ -66,14 +75,24 @@ namespace SelXPressApi.Controllers
         }
 
         /// <summary>
-        /// Create a new AttributeData item.
+        /// Method to create attribute data
         /// </summary>
+        /// <param name="attributeData"></param>
+        /// <returns></returns>
+        /// <exception cref="ForbiddenRequestException"></exception>
+        /// <exception cref="BadRequestException"></exception>
         [HttpPost]
         [ProducesResponseType(201)]
         [ProducesResponseType(400, Type = typeof(BadRequestErrorTemplate))]
+        [ProducesResponseType(401, Type = typeof(UnauthorizedErrorTemplate))]
+        [ProducesResponseType(403, Type = typeof(ForbiddenErrorTemplate))]
         [ProducesResponseType(500, Type = typeof(InternalServerErrorTemplate))]
         public async Task<IActionResult> CreateAttributeData([FromBody] CreateAttributeDataDTO attributeData)
         {
+            await _authorizationMiddleware.CheckIfTokenExists(HttpContext);
+            if (!await _authorizationMiddleware.CheckRoleIfAdmin(HttpContext))
+                throw new ForbiddenRequestException("You are not authorized to do this operation", "ATD-2001");
+            
             if (attributeData == null || attributeData.Value == null || attributeData.Key == null)
                 throw new BadRequestException("There are missing fields, please try again with some data", "ATD-1102");
 
@@ -82,37 +101,64 @@ namespace SelXPressApi.Controllers
         }
 
         /// <summary>
-        /// Update an existing AttributeData item.
+        /// Method to update the attribute data based on the id
         /// </summary>
+        /// <param name="id"></param>
+        /// <param name="attributeData"></param>
+        /// <returns></returns>
+        /// <exception cref="ForbiddenRequestException"></exception>
+        /// <exception cref="BadRequestException"></exception>
+        /// <exception cref="NotFoundException"></exception>
         [HttpPut("{id}")]
         [ProducesResponseType(200)]
         [ProducesResponseType(400, Type = typeof(BadRequestErrorTemplate))]
+        [ProducesResponseType(401, Type = typeof(UnauthorizedErrorTemplate))]
+        [ProducesResponseType(403, Type = typeof(ForbiddenErrorTemplate))]
         [ProducesResponseType(404, Type = typeof(NotFoundErrorTemplate))]
         [ProducesResponseType(500, Type = typeof(InternalServerErrorTemplate))]
         public async Task<IActionResult> UpdateAttributeData(int id, [FromBody] UpdateAttributeDataDTO attributeData)
         {
+            await _authorizationMiddleware.CheckIfTokenExists(HttpContext);
+            if (!await _authorizationMiddleware.CheckRoleIfAdmin(HttpContext))
+                throw new ForbiddenRequestException("You are not authorized to do this operation", "ATD-2001");
+            
             if (!ModelState.IsValid)
                 throw new BadRequestException("The model is wrong, a bad request occured", "ATD-1101");
+            
             if(attributeData == null)
                 throw new BadRequestException("There are missing fields, please try again with some data", "ATD-1102");
+            
             if(!await _attributeDataRepository.AttributeDataExists(id))
-                throw new NotFoundException("The AttributeData with the id : " + id + " doesn't exist", "RLE-1402");
+                throw new NotFoundException("The AttributeData with the id : " + id + " doesn't exist", "ATD-1402");
+            
             await _attributeDataRepository.UpdateAttributeData(id, attributeData);
             return Ok();
         }
 
         /// <summary>
-        /// Delete an AttributeData item by ID.
+        /// Method to delete the attribute data based on the id
         /// </summary>
+        /// <param name="id"></param>
+        /// <returns></returns>
+        /// <exception cref="ForbiddenRequestException"></exception>
+        /// <exception cref="NotFoundException"></exception>
+        /// <exception cref="BadRequestException"></exception>
         [HttpDelete("{id}")]
         [ProducesResponseType(200)]
         [ProducesResponseType(400, Type = typeof(BadRequestErrorTemplate))]
+        [ProducesResponseType(401, Type = typeof(UnauthorizedErrorTemplate))]
+        [ProducesResponseType(403, Type = typeof(ForbiddenErrorTemplate))]
         [ProducesResponseType(404, Type = typeof(NotFoundErrorTemplate))]
         [ProducesResponseType(500, Type = typeof(InternalServerErrorTemplate))]
         public async Task<IActionResult> DeleteAttributeData(int id)
         {
+            await _authorizationMiddleware.CheckIfTokenExists(HttpContext);
+            if (!await _authorizationMiddleware.CheckRoleIfAdmin(HttpContext))
+                throw new ForbiddenRequestException("You are not authorized to do this operation", "ATD-2001");
+            
             if(!await _attributeDataRepository.AttributeDataExists(id))
-                throw new NotFoundException("The AttributeData with the id : " + id + " doesn't exist", "RLE-1402");
+                throw new NotFoundException("The AttributeData with the id : " + id + " doesn't exist", "ATD-1402");
+            
             if (!ModelState.IsValid)
                 throw new BadRequestException("The model is wrong, a bad request occured", "ATD-1101");
             await _attributeDataRepository.DeleteAttributeData(id);
